refactor: split API module by feature for better structure
